+++ nusamai/src/sink/shapefile/null_shape.rs
